refactor: fix architecture layer violations (findings 138, 143)
- Extract branchProtectionStore from branchProtectionReporter (finding 138):
  persistence functions (persistReport, loadReport) moved to
  src/persistence/branchProtectionStore.ts, separating I/O from report
  generation logic
- Move validateCliPath to src/validation/cliPath.ts (finding 143):
  breaks coupling between workflows/codeMachineRunner and
  adapters/codemachine/types
