+++ src/workflows/branchProtectionReporter.ts
@@ -10,52 +10,11 @@ import * as fs from 'node:fs/promises';
 import * as path from 'node:path';
 import { z } from 'zod';
 import type { BranchProtectionCompliance } from '../adapters/github/branchProtection';
+import type { BranchProtectionReport } from '../core/models/BranchProtectionReport';
 import { validateOrThrow } from '../validation/helpers.js';
 
-/**
- * Branch protection report schema for persistence
- */
-export const BranchProtectionReportSchema = z.object({
-  schema_version: z.string(),
-  feature_id: z.string(),
-  branch: z.string(),
-  sha: z.string(),
-  base_sha: z.string(),
-  pull_number: z.number().optional(),
-  protected: z.boolean(),
-  compliant: z.boolean(),
-  required_checks: z.array(z.string()),
-  checks_passing: z.boolean(),
-  failing_checks: z.array(z.string()),
-  reviews_required: z.number(),
-  reviews_count: z.number(),
-  reviews_satisfied: z.boolean(),
-  up_to_date: z.boolean(),
-  stale_commit: z.boolean(),
-  allows_auto_merge: z.boolean(),
-  allows_force_push: z.boolean(),
-  blockers: z.array(z.string()),
-  evaluated_at: z.string(),
-  validation_mismatch: z
-    .object({
-      missing_in_registry: z.array(z.string()),
-      extra_in_registry: z.array(z.string()),
-      recommendations: z.array(z.string()),
-    })
-    .optional(),
-  metadata: z
-    .object({
-      owner: z.string(),
-      repo: z.string(),
-      protection_enabled: z.boolean(),
-      enforce_admins: z.boolean().optional(),
-      required_linear_history: z.boolean().optional(),
-      required_conversation_resolution: z.boolean().optional(),
-    })
-    .optional(),
-});
-
-export type BranchProtectionReport = z.infer<typeof BranchProtectionReportSchema>;
+export { BranchProtectionReportSchema } from '../core/models/BranchProtectionReport';
+export type { BranchProtectionReport } from '../core/models/BranchProtectionReport';
 
 /**
  * Summary for CLI display
@@ -107,7 +66,6 @@ const CommandsDataSchema = z.object({
 });
 
 const SCHEMA_VERSION = '1.0.0';
-const BRANCH_PROTECTION_FILE = 'branch_protection.json';
 
 /**
  * Generate branch protection report from compliance result
@@ -156,44 +114,6 @@ export function generateReport(
         }
       : undefined,
   };
-}
-
-/**
- * Persist branch protection report to run directory
- */
-export async function persistReport(
-  runDir: string,
-  report: BranchProtectionReport
-): Promise<string> {
-  const reportPath = path.join(runDir, 'status', BRANCH_PROTECTION_FILE);
-
-  await fs.mkdir(path.dirname(reportPath), { recursive: true });
-
-  // Validate report schema
-  const validated = BranchProtectionReportSchema.parse(report);
-
-  // Write report
-  await fs.writeFile(reportPath, JSON.stringify(validated, null, 2), 'utf-8');
-
-  return reportPath;
-}
-
-/**
- * Load branch protection report from run directory
- */
-export async function loadReport(runDir: string): Promise<BranchProtectionReport | null> {
-  const reportPath = path.join(runDir, 'status', BRANCH_PROTECTION_FILE);
-
-  try {
-    const content = await fs.readFile(reportPath, 'utf-8');
-    const data: unknown = JSON.parse(content);
-    return BranchProtectionReportSchema.parse(data);
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code === 'ENOENT') {
-      return null;
-    }
-    throw error;
-  }
 }
 
 /**
