+++ todos/debt/149-complete-high-god-function-doctor-run-158-lines-deeply-b4c64271.md
@@ -1,7 +1,7 @@
 # God Function Doctor run 158 Lines Deeply Nested Telemetry Conditionals
 
 **ID:** 149
-**Status:** pending
+**Status:** complete
 **Severity:** high
 **Category:** complexity
 **Effort:** medium
