+++ todos/debt/153-complete-high-handlefailure-method-6-return-paths-mixe-9045008c.md
@@ -1,7 +1,7 @@
 # handleFailure Method 6 Return Paths Mixed Retry Artifact Metric Concerns
 
 **ID:** 153
-**Status:** pending
+**Status:** complete
 **Severity:** high
 **Category:** complexity
 **Effort:** medium
