+++ todos/debt/156-complete-medium-checkrunstatus-switch-statement-mixes-st-94df4ce5.md
@@ -1,7 +1,7 @@
 # checkRunStatus Switch Statement Mixes State Machine Logic with Inline Diagnostic Mutation
 
 **ID:** 156
-**Status:** pending
+**Status:** complete
 **Severity:** medium
 **Category:** complexity
 **Effort:** small
