+++ todos/debt/157-complete-medium-composespecification-god-function-180-li-4b5b7706.md
@@ -1,7 +1,7 @@
 # composeSpecification God Function 180 Lines 12 Numbered Sequential Steps
 
 **ID:** 157
-**Status:** pending
+**Status:** complete
 **Severity:** medium
 **Category:** complexity
 **Effort:** medium
