+++ todos/debt/166-complete-medium-analyzeresumestate-orchestrates-7-sub-ch-aa65b3df.md
@@ -1,7 +1,7 @@
 # analyzeResumeState Orchestrates 7 Sub-Checks Without Separating Optional vs Mandatory
 
 **ID:** 166
-**Status:** pending
+**Status:** complete
 **Severity:** medium
 **Category:** complexity
 **Effort:** small
